Screening tool builder (#430)
* basic migrations

* update schema based on prd discussion

* add basic screening tool model and tests

* remove a typing workaround since objection has been updated since the reqd pr

* just adding the patient-screening-tool-submission model file...not done

* mostly finish PatientScreeningToolSubmission model...now needs specs

* spell "score" correctly like a person who can type words

* mostly add tests, still kinda broken

* that test loses, i win

* add migration for screening tool score range

* add screening-tool-score-range model (needs tests)

* update migrations

* fix indexes and add tests

* start screening-tool-resolver

* add types

* this test was a lie

* screening tool resolver specs

* add screening tool score range resolver

* update schema for screening tool score range resolver

* prettier and linting

* tests for screening tool score range resolver

* committing just in case

* patient-screening-tool-resolver done missing types and tests

* typings and prettier

* fix dumb things

* patient screening tool submission resolver specs

* starting

* more

* getting there

* display

* more prettier

* do not edit risk area associated with screening tool for now

* add score ranges

* more prettier

* linting

* create tools/score ranges/suggestions

* more prettier

* more changes

* really close with admin...just one little bug left

* prettier

* lintastic

* boom

* regen documents.json

* more linting fixes

* fix existing broken tests

* spelling fix

* doh

* add snapshot tests

* finish adding new tests

* prettier

* some code review fixes

* typing fixes

* add a couple of tests for care-plan-suggestion-create functions

* concern suggestions working...goal suggestions tbd

* prettier

* goal suggestions are good to go now

* typings on typings on typings

* fix busted css

* snapshot ugh

+++ app/builder-container/question-condition-row.tsx
@@ -17,7 +17,7 @@ export interface IDeleteOptions { variables: questionConditionDeleteMutationVari
 interface IProps {
   answer?: FullAnswerFragment;
   questionCondition: FullQuestionConditionFragment;
-  deleteQuestionCondition: (
+  deleteQuestionCondition?: (
     options: IDeleteOptions,
   ) => { data: { questionConditionDelete: FullQuestionConditionFragment } };
 }
@@ -31,7 +31,10 @@ class QuestionConditionRow extends React.Component<IProps> {
 
   onClick() {
     const { deleteQuestionCondition, questionCondition } = this.props;
-    deleteQuestionCondition({ variables: { questionConditionId: questionCondition.id } });
+
+    if (deleteQuestionCondition) {
+      deleteQuestionCondition({ variables: { questionConditionId: questionCondition.id } });
+    }
   }
 
   render() {
@@ -62,8 +65,8 @@ export default compose(
     name: 'deleteQuestionCondition',
     options: {
       refetchQueries: [
-        'getQuestionsForRiskArea',
+        'getQuestionsForRiskAreaOrScreeningTool',
       ],
     },
   }),
-)(QuestionConditionRow as any) as any;
+)(QuestionConditionRow);
